requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.0
annotated-types==0.7.0
anyascii==0.3.2
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.3.0
audioread==3.0.1
babel==2.16.0
bangla==0.0.2
beautifulsoup4==4.12.3
blinker==1.9.0
blis==1.1.0
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.10
Cython==3.0.11
dateparser==1.1.8
decorator==5.1.1
deepface==0.0.93
docopt==0.6.2
einops==0.8.0
encodec==0.1.1
filelock==3.16.1
fire==0.7.0
Flask==2.2.2
Flask-Cors==5.0.0
fonttools==4.55.3
frozenlist==1.5.0
fsspec==2024.12.0
g2pkk==0.1.2
gast==0.4.0
gdown==5.2.0
google-auth==2.37.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.69.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.1
gruut-lang-en==2.0.1
gruut-lang-es==2.0.1
gruut-lang-fr==2.0.2
gunicorn==23.0.0
h5py==3.12.1
hangul-romanize==0.1.0
huggingface-hub==0.27.1
idna==3.10
inflect==7.5.0
itsdangerous==2.2.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.5
joblib==1.4.2
jsonlines==1.2.0
keras==2.14.0
kiwisolver==1.4.7
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.0
llvmlite==0.43.0
lz4==4.3.3
Mako==1.3.8
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.2.0
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
mtcnn==0.1.1
multidict==6.1.0
murmurhash==1.0.11
namex==0.0.8
networkx==2.8.8
nltk==3.9.1
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optree==0.13.1
packaging==24.2
pandas==1.5.3
pillow==11.1.0
platformdirs==4.3.6
pooch==1.8.2
preshed==3.0.9
propcache==0.2.1
protobuf==4.25.5
psutil==6.1.1
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
pypinyin==0.53.0
pysbd==0.3.4
PySocks==1.7.1
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
retina-face==0.0.17
rich==13.9.4
rsa==4.9
safetensors==0.5.2
scikit-learn==1.6.0
scipy==1.11.4
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
soundfile==0.13.0
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.8.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.37
srsly==2.5.0
SudachiDict-core==20241021
SudachiPy==0.6.9
sympy==1.13.1
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
thinc==8.3.3
threadpoolctl==3.5.0
tokenizers==0.21.0
tqdm==4.67.1
trainer==0.0.36
transformers==4.47.1
triton==2.1.0
typeguard==4.4.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
ultralytics==8.3.58
ultralytics-thop==2.0.13
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.3.0
wasabi==1.1.3
weasel==0.4.1
wrapt==1.14.1
yarl==1.18.3
Flask-SocketIO==5.3.3
eventlet==0.33.0
psycopg2-binary==2.9.3
Werkzeug==2.2.3
tf-keras
dnspython==2.2.1
